template/backend/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
av==14.4.0
certifi==2025.4.26
cffi==1.17.1
click==8.2.1
colorama==0.4.6
distro==1.9.0
docstring_parser==0.16
dotenv==0.9.9
eval_type_backport==0.2.2
frozenlist==1.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
livekit==1.0.9
livekit-agents==1.1.1
livekit-api==1.0.2
livekit-plugins-hedra==1.1.1
livekit-plugins-openai==1.1.1
livekit-plugins-elevenlabs==1.1.1
livekit-plugins-silero==1.1.1
livekit-protocol==1.0.3
multidict==6.4.4
nest-asyncio==1.6.0
numpy==2.3.0
openai==1.86.0
pillow==11.2.1
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
PyJWT==2.10.1
python-dotenv==1.1.0
sniffio==1.3.1
sounddevice==0.5.2
tqdm==4.67.1
types-protobuf==4.25.0.20240417
typing-inspection==0.4.1
typing_extensions==4.14.0
watchfiles==1.0.5
websockets==15.0.1
yarl==1.20.1
